Convert Plugin install to only build once (#1708)
When running in the OpenSearch distribution the bundled jdk version is
used, this has a more dramatic impact on the capibilities and
limitations of plugins.  Removing the extra versions as it was providing
a false send of test coverage.  These test confirm basic installation
and service startup - no depth of coverage that would be exposed by
runtime versions.

Signed-off-by: Peter Nied <petern@amazon.com>